✨ 内容 - Todo 💫 修复 当删除多余扩展包时 残留的Git生成文件会导致无限加载

+++ Editor.GUI.CLI/External/AIO.Git/MenuItem.Git.cs
@@ -19,30 +19,25 @@ namespace AIO.UEditor
         /// 生成
         /// </summary>
         [MenuItem("Git/~~~ Generate ~~~", false, 9999)]
-        [AInit(mode: EInitAttrMode.Editor, int.MaxValue - 1)]
+        [AInit(mode: EInitAttrMode.Editor, ushort.MaxValue - 1)]
         internal static void Generate()
         {
             Debug.Log($"<b><color=#5DADE2>[GIT]</color></b> {CMD_GIT} Generate");
 
-            var packageInfos = AssetDatabase.FindAssets("package", new string[] { "Packages" })
-                .Select(AssetDatabase.GUIDToAssetPath)
-                .Where(x => AssetDatabase.LoadAssetAtPath<TextAsset>(x) != null)
-                .Select(PackageInfo.FindForAssetPath)
-                .GroupBy(x => x.assetPath)
-                .Select(x => x.First())
-                .Where(x =>
+            var packageInfos = AssetDatabase.FindAssets("package", new string[] { "Packages" }).
+                Select(AssetDatabase.GUIDToAssetPath).Where(x => AssetDatabase.LoadAssetAtPath<TextAsset>(x) != null).
+                Select(PackageInfo.FindForAssetPath).GroupBy(x => x.assetPath).Select(x => x.First()).Where(x =>
                     File.Exists(Path.Combine(EHelper.Path.Project, x.resolvedPath, ".git")) ||
                     Directory.Exists(Path.Combine(EHelper.Path.Project, x.resolvedPath, ".git"))
-                )
-                .ToList();
+                ).ToList();
 
             var change = CreateProject();
             if (CreateTemplate(packageInfos)) change = true;
             if (!change) return;
 
             AssetDatabase.Refresh();
-            var RefreshSettings = typeof(AssetDatabase)
-                .GetMethod("RefreshSettings", BindingFlags.Static | BindingFlags.Public);
+            var RefreshSettings =
+                typeof(AssetDatabase).GetMethod("RefreshSettings", BindingFlags.Static | BindingFlags.Public);
             if (RefreshSettings != null) RefreshSettings.Invoke(null, null);
             CompilationPipeline.RequestScriptCompilation();
         }
